+++ src/main/java/com/sprint/monew/domain/activity/UserActivityDto.java
@@ -2,8 +2,7 @@ package com.sprint.monew.domain.activity;
 
 import com.sprint.monew.domain.article.dto.ArticleViewDto;
 import com.sprint.monew.domain.comment.dto.CommentDto;
-import com.sprint.monew.domain.interest.dto.SubscriptionDto;
-
+import com.sprint.monew.domain.interest.subscription.SubscriptionDto;
 import java.time.Instant;
 import java.util.List;
 import java.util.UUID;
