+++ src/main/java/com/sprint/monew/domain/activity/UserActivityQueryRepository.java
@@ -10,8 +10,8 @@ import com.sprint.monew.domain.article.dto.ArticleViewDto;
 import com.sprint.monew.domain.comment.QComment;
 import com.sprint.monew.domain.comment.dto.CommentDto;
 import com.sprint.monew.domain.interest.QInterest;
-import com.sprint.monew.domain.interest.dto.SubscriptionDto;
 import com.sprint.monew.domain.interest.subscription.QSubscription;
+import com.sprint.monew.domain.interest.subscription.SubscriptionDto;
 import com.sprint.monew.domain.like.QLike;
 import com.sprint.monew.domain.user.QUser;
 import com.sprint.monew.domain.user.User;
